price_model/requirements.txt
python-geohash==0.8.5
tensorflow==2.18.0
keras==3.8.0
pandas==2.2.2
matplotlib==3.8.3
seaborn==0.13.2
requests==2.31.0
pika==1.3.1
numpy==1.26.2
scikit-learn==1.3.2
kneed==0.8.5
folium==0.14.0

